feat: add InputNricSection (#25)
* feat: add InputNricSection to be used whenever we need NRIC input/scanning

* fix: failing tests

* chore: remove unused styles

* chore: remove ! assertion
